Move search test to it's own test file, uncomment most of the tests
